Refactor(PostgresProvider, CitusProvider): Pull out Postgres Table Creation: Created and extended from TableCreator & TableQueryGenerator, removed statement executor
