+++ src/sqlancer/postgres/PostgresProvider.java
@@ -10,38 +10,13 @@ import java.util.Arrays;
 
 import com.google.auto.service.AutoService;
 
-import sqlancer.AbstractAction;
 import sqlancer.DatabaseProvider;
-import sqlancer.IgnoreMeException;
 import sqlancer.MainOptions;
 import sqlancer.Randomly;
 import sqlancer.SQLConnection;
 import sqlancer.SQLProviderAdapter;
-import sqlancer.StatementExecutor;
-import sqlancer.common.DBMSCommon;
 import sqlancer.common.query.SQLQueryAdapter;
-import sqlancer.common.query.SQLQueryProvider;
 import sqlancer.common.query.SQLancerResultSet;
-import sqlancer.postgres.gen.PostgresAlterTableGenerator;
-import sqlancer.postgres.gen.PostgresAnalyzeGenerator;
-import sqlancer.postgres.gen.PostgresClusterGenerator;
-import sqlancer.postgres.gen.PostgresCommentGenerator;
-import sqlancer.postgres.gen.PostgresDeleteGenerator;
-import sqlancer.postgres.gen.PostgresDiscardGenerator;
-import sqlancer.postgres.gen.PostgresDropIndexGenerator;
-import sqlancer.postgres.gen.PostgresIndexGenerator;
-import sqlancer.postgres.gen.PostgresInsertGenerator;
-import sqlancer.postgres.gen.PostgresNotifyGenerator;
-import sqlancer.postgres.gen.PostgresReindexGenerator;
-import sqlancer.postgres.gen.PostgresSequenceGenerator;
-import sqlancer.postgres.gen.PostgresSetGenerator;
-import sqlancer.postgres.gen.PostgresStatisticsGenerator;
-import sqlancer.postgres.gen.PostgresTableGenerator;
-import sqlancer.postgres.gen.PostgresTransactionGenerator;
-import sqlancer.postgres.gen.PostgresTruncateGenerator;
-import sqlancer.postgres.gen.PostgresUpdateGenerator;
-import sqlancer.postgres.gen.PostgresVacuumGenerator;
-import sqlancer.postgres.gen.PostgresViewGenerator;
 
 // EXISTS
 // IN
@@ -72,129 +47,11 @@ public class PostgresProvider extends SQLProviderAdapter<PostgresGlobalState, Po
         super(globalClass, optionClass);
     }
 
-    public enum Action implements AbstractAction<PostgresGlobalState> {
-        ANALYZE(PostgresAnalyzeGenerator::create), //
-        ALTER_TABLE(g -> PostgresAlterTableGenerator.create(g.getSchema().getRandomTable(t -> !t.isView()), g,
-                generateOnlyKnown)), //
-        CLUSTER(PostgresClusterGenerator::create), //
-        COMMIT(g -> {
-            SQLQueryAdapter query;
-            if (Randomly.getBoolean()) {
-                query = new SQLQueryAdapter("COMMIT", true);
-            } else if (Randomly.getBoolean()) {
-                query = PostgresTransactionGenerator.executeBegin();
-            } else {
-                query = new SQLQueryAdapter("ROLLBACK", true);
-            }
-            return query;
-        }), //
-        CREATE_STATISTICS(PostgresStatisticsGenerator::insert), //
-        DROP_STATISTICS(PostgresStatisticsGenerator::remove), //
-        DELETE(PostgresDeleteGenerator::create), //
-        DISCARD(PostgresDiscardGenerator::create), //
-        DROP_INDEX(PostgresDropIndexGenerator::create), //
-        INSERT(PostgresInsertGenerator::insert), //
-        UPDATE(PostgresUpdateGenerator::create), //
-        TRUNCATE(PostgresTruncateGenerator::create), //
-        VACUUM(PostgresVacuumGenerator::create), //
-        REINDEX(PostgresReindexGenerator::create), //
-        SET(PostgresSetGenerator::create), //
-        CREATE_INDEX(PostgresIndexGenerator::generate), //
-        SET_CONSTRAINTS((g) -> {
-            StringBuilder sb = new StringBuilder();
-            sb.append("SET CONSTRAINTS ALL ");
-            sb.append(Randomly.fromOptions("DEFERRED", "IMMEDIATE"));
-            return new SQLQueryAdapter(sb.toString());
-        }), //
-        RESET_ROLE((g) -> new SQLQueryAdapter("RESET ROLE")), //
-        COMMENT_ON(PostgresCommentGenerator::generate), //
-        RESET((g) -> new SQLQueryAdapter("RESET ALL") /*
-                                                       * https://www.postgresql.org/docs/devel/sql-reset.html TODO: also
-                                                       * configuration parameter
-                                                       */), //
-        NOTIFY(PostgresNotifyGenerator::createNotify), //
-        LISTEN((g) -> PostgresNotifyGenerator.createListen()), //
-        UNLISTEN((g) -> PostgresNotifyGenerator.createUnlisten()), //
-        CREATE_SEQUENCE(PostgresSequenceGenerator::createSequence), //
-        CREATE_VIEW(PostgresViewGenerator::create);
-
-        private final SQLQueryProvider<PostgresGlobalState> sqlQueryProvider;
-
-        Action(SQLQueryProvider<PostgresGlobalState> sqlQueryProvider) {
-            this.sqlQueryProvider = sqlQueryProvider;
-        }
-
-        @Override
-        public SQLQueryAdapter getQuery(PostgresGlobalState state) throws Exception {
-            return sqlQueryProvider.getQuery(state);
-        }
-    }
-
-    public static int mapActions(PostgresGlobalState globalState, Action a) {
-        Randomly r = globalState.getRandomly();
-        int nrPerformed;
-        switch (a) {
-        case CREATE_INDEX:
-        case CLUSTER:
-            nrPerformed = r.getInteger(0, 3);
-            break;
-        case CREATE_STATISTICS:
-            nrPerformed = r.getInteger(0, 5);
-            break;
-        case DISCARD:
-        case DROP_INDEX:
-            nrPerformed = r.getInteger(0, 5);
-            break;
-        case COMMIT:
-            nrPerformed = r.getInteger(0, 0);
-            break;
-        case ALTER_TABLE:
-            nrPerformed = r.getInteger(0, 5);
-            break;
-        case REINDEX:
-        case RESET:
-            nrPerformed = r.getInteger(0, 3);
-            break;
-        case DELETE:
-        case RESET_ROLE:
-        case SET:
-            nrPerformed = r.getInteger(0, 5);
-            break;
-        case ANALYZE:
-            nrPerformed = r.getInteger(0, 3);
-            break;
-        case VACUUM:
-        case SET_CONSTRAINTS:
-        case COMMENT_ON:
-        case NOTIFY:
-        case LISTEN:
-        case UNLISTEN:
-        case CREATE_SEQUENCE:
-        case DROP_STATISTICS:
-        case TRUNCATE:
-            nrPerformed = r.getInteger(0, 2);
-            break;
-        case CREATE_VIEW:
-            nrPerformed = r.getInteger(0, 2);
-            break;
-        case UPDATE:
-            nrPerformed = r.getInteger(0, 10);
-            break;
-        case INSERT:
-            nrPerformed = r.getInteger(0, globalState.getOptions().getMaxNumberInserts());
-            break;
-        default:
-            throw new AssertionError(a);
-        }
-        return nrPerformed;
-
-    }
-
     @Override
     public void generateDatabase(PostgresGlobalState globalState) throws Exception {
         readFunctions(globalState);
-        createTables(globalState, Randomly.fromOptions(4, 5, 6));
-        prepareTables(globalState);
+        PostgresTableCreator tableCreator = new PostgresTableCreator(globalState, generateOnlyKnown);
+        tableCreator.create();
 
         extensionsList = globalState.getDbmsSpecificOptions().extensions;
         if (!extensionsList.isEmpty()) {
@@ -294,31 +151,6 @@ public class PostgresProvider extends SQLProviderAdapter<PostgresGlobalState, Po
             Character functionType = rs.getString(2).charAt(0);
             globalState.addFunctionAndType(functionName, functionType);
         }
-    }
-
-    protected void createTables(PostgresGlobalState globalState, int numTables) throws Exception {
-        while (globalState.getSchema().getDatabaseTables().size() < numTables) {
-            try {
-                String tableName = DBMSCommon.createTableName(globalState.getSchema().getDatabaseTables().size());
-                SQLQueryAdapter createTable = PostgresTableGenerator.generate(tableName, globalState.getSchema(),
-                        generateOnlyKnown, globalState);
-                globalState.executeStatement(createTable);
-            } catch (IgnoreMeException e) {
-
-            }
-        }
-    }
-
-    protected void prepareTables(PostgresGlobalState globalState) throws Exception {
-        StatementExecutor<PostgresGlobalState, Action> se = new StatementExecutor<>(globalState, Action.values(),
-                PostgresProvider::mapActions, (q) -> {
-                    if (globalState.getSchema().getDatabaseTables().isEmpty()) {
-                        throw new IgnoreMeException();
-                    }
-                });
-        se.executeStatements();
-        globalState.executeStatement(new SQLQueryAdapter("COMMIT", true));
-        globalState.executeStatement(new SQLQueryAdapter("SET SESSION statement_timeout = 5000;\n"));
     }
 
     private String getCreateDatabaseCommand(PostgresGlobalState state) {
